--- python/dask-raft/raft/test/test_raft.py
@@ -1,31 +1,0 @@
-# Copyright (c) 2020-2022, NVIDIA CORPORATION.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-import pytest
-import sys
-
-try:
-    import raft
-except ImportError:
-    print("Skipping RAFT tests")
-    pytestmart = pytest.mark.skip
-
-pytestmark = pytest.mark.skipif(
-    'raft' not in sys.argv, reason="marker to allow integration of RAFT"
-)
-
-
-def test_raft():
-    assert raft.raft_include_test()
