+++ python/pylibraft/pylibraft/common/CMakeLists.txt
@@ -13,7 +13,7 @@
 # =============================================================================
 
 # Set the list of Cython files to build
-set(cython_sources cuda.pyx handle.pyx)
+set(cython_sources cuda.pyx handle.pyx interruptible.pyx)
 set(linked_libraries raft::raft)
 
 # Build all of the Cython targets
