+++ python/pylibraft/pyproject.toml
@@ -20,5 +20,5 @@ requires = [
     "cython>=0.29,<0.30",
     "scikit-build>=0.13.1",
     "cmake>=3.20.1,!=3.23.0",
-    "ninja",
+    "ninja"
 ]
