+++ python/raft-dask/pyproject.toml
@@ -1,5 +1,4 @@
-#
-# Copyright (c) 2019-2022, NVIDIA CORPORATION.
+# Copyright (c) 2022, NVIDIA CORPORATION.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,11 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 
-from cuda.ccudart cimport cudaStream_t
+[build-system]
 
-cdef class Stream:
-    cdef cudaStream_t s
-
-    cdef cudaStream_t getStream(self)
+requires = [
+    "wheel",
+    "setuptools",
+    "cython>=0.29,<0.30",
+    "scikit-build>=0.13.1",
+    "cmake>=3.20.1,!=3.23.0",
+    "ninja",
+    "pylibraft"
+]
