+++ python/raft-dask/raft_dask/common/CMakeLists.txt
@@ -12,8 +12,11 @@
 # the License.
 # =============================================================================
 
-set(cython_sources cuda.pyx handle.pyx interruptible.pyx)
-set(linked_libraries raft::raft)
+include(${raft-dask-python_SOURCE_DIR}/cmake/thirdparty/get_nccl.cmake)
+find_package(ucx REQUIRED)
+
+set(cython_sources comms_utils.pyx nccl.pyx)
+set(linked_libraries raft::raft NCCL::NCCL ucx::ucp)
 rapids_cython_create_modules(
     SOURCE_FILES "${cython_sources}"
     LINKED_LIBRARIES "${linked_libraries}"
