+++ python/raft-dask/setup.cfg
@@ -9,10 +9,10 @@ exclude = __init__.py,versioneer.py
 [versioneer]
 VCS = git
 style = pep440
-versionfile_source = raft/_version.py
-versionfile_build = raft/_version.py
+versionfile_source = raft_dask/_version.py
+versionfile_build = raft_dask/_version.py
 tag_prefix = v
-parentdir_prefix = raft-
+parentdir_prefix = raft_dask-
 
 [isort]
 line_length=79
@@ -33,7 +33,7 @@ known_rapids=
     dask_cudf
     rmm
 known_first_party=
-    raft
+    raft_dask
 default_section=THIRDPARTY
 sections=FUTURE,STDLIB,THIRDPARTY,DASK,RAPIDS,FIRSTPARTY,LOCALFOLDER
 skip=
